Bug fixed

+++ price.py
@@ -1,5 +1,5 @@
 def format_price(price):
-    price=int(price)
+    price=float(price)
     return f'Цена: {price} руб.'
 
 result = format_price(56.24)
